Updated database design, exceptions

+++ ecommerce-backend/src/main/java/com/nashtech/rookies/ecommerce/services/cart/impl/CartItemServiceImpl.java
@@ -3,12 +3,13 @@ package com.nashtech.rookies.ecommerce.services.cart.impl;
 import java.util.ArrayList;
 import java.util.List;
 
+import com.nashtech.rookies.ecommerce.models.cart.Cart;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.nashtech.rookies.ecommerce.dto.cart.requests.CartItemRequestDTO;
 import com.nashtech.rookies.ecommerce.dto.cart.responses.CartItemResponseDTO;
-import com.nashtech.rookies.ecommerce.exceptions.ResourceNotFoundException;
+import com.nashtech.rookies.ecommerce.handlers.exceptions.NotFoundException;
 import com.nashtech.rookies.ecommerce.mappers.cart.CartItemMapper;
 import com.nashtech.rookies.ecommerce.models.cart.CartItem;
 import com.nashtech.rookies.ecommerce.repositories.cart.CartItemRepository;
@@ -39,21 +40,20 @@ public class CartItemServiceImpl extends CommonServiceImpl<CartItem, Long> imple
     public CartItemResponseDTO createCartItem(CartItemRequestDTO cartItemRequestDTO) {
         CartItem cartItem = new CartItem();
         if (cartRepository.existsById(cartItemRequestDTO.cartId())) {
-            cartItem.setCart(cartRepository.findById(cartItemRequestDTO.cartId()).get());
+            Cart cart = cartRepository.findById(cartItemRequestDTO.cartId()).get();
+            cartItem.setCart(cart);
             if (productRepository.existsById(cartItemRequestDTO.productId())) {
                 cartItem.setProduct(productRepository.findById(cartItemRequestDTO.productId()).get());
                 cartItem.setQuantity(cartItemRequestDTO.quantity());
                 cartItem = cartItemRepository.saveAndFlush(cartItem);
+                cart.setQuantity(cart.getQuantity() + 1L);
+                cartRepository.saveAndFlush(cart);
                 return cartItemMapper.toResponseDTO(cartItem);
             } else
-                throw new ResourceNotFoundException("Not found Product with an id: " + cartItemRequestDTO.productId());
+                throw new NotFoundException("Not found Product with an id: " + cartItemRequestDTO.productId());
         } else
-            throw new ResourceNotFoundException("Not found Cart with an id: " + cartItemRequestDTO.cartId());
+            throw new NotFoundException("Not found Cart with an id: " + cartItemRequestDTO.cartId());
     }
-
-//    new CartItemResponseDTO(cartItem.getId(), cartItem.getCreatedOn(), cartItem.getLastUpdatedOn(),
-//                            cartItem.getQuantity(), cartItem.getCart().getId(), cartItem.getOrder().getId(),
-//                            cartItem.getProduct().getId())
 
     @Override
     public List<CartItemResponseDTO> getCartItem() {
@@ -71,7 +71,7 @@ public class CartItemServiceImpl extends CommonServiceImpl<CartItem, Long> imple
             cartItemResponseDTOs.add(cartItemMapper.toResponseDTO(cartItem));
             return cartItemResponseDTOs;
         } else
-            throw new ResourceNotFoundException("Not found a Cart Item with an id: " + id);
+            throw new NotFoundException("Not found a Cart Item with an id: " + id);
     }
 
     @Override
@@ -82,13 +82,21 @@ public class CartItemServiceImpl extends CommonServiceImpl<CartItem, Long> imple
             cartItem = cartItemRepository.saveAndFlush(cartItem);
             return cartItemMapper.toResponseDTO(cartItem);
         } else
-            throw new ResourceNotFoundException("Not found a Cart Item with an id: " + id);
+            throw new NotFoundException("Not found a Cart Item with an id: " + id);
     }
 
     @Override
     @Transactional
     public String deleteCartItem(Long id) {
-        cartItemRepository.deleteById(id);
-        return "Delete Cart Item successful";
+        if (cartItemRepository.existsById(id)) {
+            CartItem cartItem = cartItemRepository.findById(id).get();
+            Cart cart = cartRepository.findById(cartItem.getCart().getId()).get();
+            cart.setQuantity(cart.getQuantity() - 1);
+            cartRepository.saveAndFlush(cart);
+            cartItemRepository.deleteById(id);
+            return "Delete successful";
+        } else {
+            throw new NotFoundException("Not found a Cart Item with an id: " + id);
+        }
     }
 }
